Expand on reviewing workflow in rationale (#9)
Don't default to reviews being optional, but allow overriding the process when needed (same effect, but different mindset: it should be exceptional to skip the reviews). Add link to code owners functionality of GitHub.

This reflects my preference on how to handle changes:
- there is no change too small for a PR (CI checks are still useful, small changes are often the ones where one is least careful about checking what one has changed thinking "it's just a simple change, what could go wrong")
- reviewing small changes is faster than reviewing big changes
- submitting all changes through PRs broadcasts to others that this particular thing has been/will be fixed rather than hiding it in a commit pushed silently to main
- the above may make others realize that other, similar "simple fixes" are required as well

The downside of this is that there is a risk of PRs stalling if the above is taken too literally. That's why it's important to:
- set a cutoff for a reasonable waiting time for reviews (I believe a week at most is reasonable; if a person wishes to review but can't within the time they should comment saying that they want to review but don't have the time)
- encourage everyone review actively, it benefits everyone

Since these may only be my own preferences, I'm not sure how much of the above rationale to put in the actual rationale.